+++ lib/config.js
@@ -56,6 +56,7 @@ const rules = [
 const template = ({
   head = '<title>mdx-deck</title>',
   body = '',
+  card,
   js,
   publicPath
 }) => `<!DOCTYPE html>
@@ -64,7 +65,12 @@ const template = ({
 <meta charset='utf-8'>
 <meta name='viewport' content='width=device-width,initial-scale=1'>
 <style>*{box-sizing:border-box}body{font-family:system-ui,sans-serif;margin:0}</style>
+<meta name='generator' content='mdx-deck'>
 ${head}
+${card ? `
+<meta name='twitter:card' content='summary_large'>
+<meta name='og:image' content='${card}'>
+` : ''}
 </head>
 <body>
 <div id=root>${body}</div>
