+++ lib/html.js
@@ -39,7 +39,7 @@ const getApp = async opts => {
         path.resolve(opts.tempdir, './App.js')
       ).default
       rimraf(opts.tempdir, err => {
-        console.error(err)
+        if (err) console.error(err)
       })
       resolve(App)
     })
