+++ lib/screenshot.js
@@ -1,3 +1,4 @@
+const fs = require('fs')
 const path = require('path')
 const puppeteer = require('puppeteer')
 
@@ -8,6 +9,9 @@ module.exports = async (opts) => {
     filename = 'card.png'
   } = opts
   const file = path.join(opts.outDir, filename)
+
+  if (!fs.existsSync(opts.outDir)) fs.mkdirSync(opts.outDir)
+
   const browser = await puppeteer.launch()
   const page = await browser.newPage()
 
@@ -30,5 +34,5 @@ module.exports = async (opts) => {
 
   await browser.close()
 
-  return file
+  return filename
 }
